Add environment variable SUDO_DISABLED for avoiding using sudo command
td-agent installer uses sudo command regardless of whether the script is
run by root user or not. This causes problem on system where sudo command
is disabled for all users. In this change we introduce an environment
variable for avoiding using sudo command when installer is run by root
user. When SUDO_DISABLED is set to 1, we don't use sudo in our own script,
and patch downloaded td-agent installer so it doesn't use sudo command when
it is run by root user.